Lot's and lot's of Debugging, fixing things and starting to actually run this
It's starting to work, the turns thing seems to be working now.
ToString is done for a bunch of stuff, did a bit of work on domestic
and everything compiles now which is nice. there's a big problem with a
null pointer exception when you try and do foreign affairs that we need
to take a look at